# docs/swagger.yaml
basePath: /
definitions:
  handler.LoginDto:
    properties:
      login:
        type: string
      senha:
        type: string
    type: object
  handler.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API para gestão de chamados
  title: Gestor de Chamados API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Autentica um usuário e retorna tokens JWT.
      parameters:
      - description: Login e senha do usuário
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/handler.LoginDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /refresh:
    post:
      consumes:
      - application/json
      description: Atualiza os tokens JWT usando um token de refresh.
      parameters:
      - description: Token de refresh
        in: body
        name: refreshRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh
      tags:
      - auth
...
swagger: "2.0"